cretae API for employee salary statement

+++ src/main/java/com/timesheetspro_api/common/dto/companyShiftDto/CompanyShiftDto.java
@@ -13,5 +13,5 @@ public class CompanyShiftDto {
     private Timestamp startTime;
     private Timestamp endTime;
     private float hours;
-    private float totalHours;
+    private Integer totalHours;
 }
